Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters

Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters

plot_gex_bivariate_qc(
  sce_list,
  x,
  y,
  color = NULL,
  shape = NULL,
  x_filters = NULL,
  y_filters = NULL,
  x_log = TRUE,
  y_log = TRUE,
  alpha = 1,
  text_size = 3,
  point_size = 0.05,
  facet_rows = NULL,
  facet_columns = NULL,
  facet_type = "grid",
  ...
)

Arguments

sce_list
list of SingleCellExperiment object

x
numeric column from colData that is in all objects in sce_list

y
numeric column from colData that is in all objects in sce_list

color
column from colData that is in all objects in sce_list

shape
column from colData that is in all objects in sce_list

x_filters
list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values

y_filters
list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values

x_log
whether to use log x-axis

y_log
whether to use log y-axis

text_size
font size for annotations

facet_rows
columns to facet on

facet_columns
columns to facet on

facet_type
either "wrap" or "grid", same as ggplot

...
other params passed into either facet_wrap or facet_grid, depending on facet_type parameter

Value

Examples

NULL

#> NULL